Struct bookworm_tui::ui::terminal_ui::AppInterface

pub(crate) struct AppInterface<D: AppDatabase + Send + Sync + 'static, B: Backend> {
    active_view: BorderWidget<D, B>,
    ui_state: UIState<D>,
    update_tui: bool,
    settings_path: Option<PathBuf>,
    event_receiver: EventStream,
    app_channel: AppChannel<D>,
    lifetime_marker: PhantomData<B>,
}

Fields
active_view: BorderWidget<D, B>ui_state: UIState<D>update_tui: boolsettings_path: Option<PathBuf>event_receiver: EventStreamapp_channel: AppChannel<D>lifetime_marker: PhantomData<B>
Implementations

impl<D: AppDatabase + Send + Sync, B: 'static + Backend> AppInterface<D, B>

pub(crate) async fn new<S: Into<String>>(
name: S,
settings: InterfaceSettings,
settings_path: Option<PathBuf>,
app_channel: AppChannel<D>,
event_receiver: EventStream,
sort_settings: SortSettings
) -> AppInterface<D, B>

Returns a new interface, instantiated with the provided settings and database.

Arguments

name - The application instance name. Not to confused with the file name.

settings - The interface settings.

settings_path - The settings path (used to persist settings).

async fn read_user_input(&mut self) -> Result<bool, TuiError<D::Error>>

Reads and handles user input. On success, returns a bool indicating whether to continue or not.

Errors

This function may error if executing a particular action fails.

fn take_update(&mut self) -> bool

pub(crate) async fn run(
&mut self,
terminal: &mut Terminal<B>
) -> Result<(), TuiError<D::Error>>

Runs the application - including handling user inputs and refreshing the output.

Arguments

terminal - The terminal to output text to.

Errors

This function will return an error if running the program fails for any reason.

async fn write_settings(&self) -> Result<(), TuiError<D::Error>>