Function bookworm_tui::ui::utils::split_chunk_into_columns

pub fn split_chunk_into_columns(chunk: Rect, num_cols: u16) -> Vec<Rect>

Splits chunk into num_cols columns with widths differing by no more than one, and adding up to the width of chunk, except when num_cols is 0. If called with sequentially increasing or decreasing values, chunk sizes will never decrease or increase, respectively.

Arguments

chunk - A chunk which the columns will be placed into.

num_cols - The number of columns to fit.