Trait bookworm_tui::ui::widgets::Widget

pub(crate) trait Widget<D: AppDatabase + Send + Sync, B: Backend> {
    fn prepare_render<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        state: &'life1 mut UIState<D>,
        chunk: Rect
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn render_into_frame(
        &self,
        f: &mut Frame<'_, B>,
        state: &UIState<D>,
        chunk: Rect
    );
    fn handle_input<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self,
        event: Event,
        state: &'life1 mut UIState<D>,
        app: &'life2 mut AppChannel<D>
    ) -> Pin<Box<dyn Future<Output = Result<ApplicationTask, TuiError<D::Error>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
}

Required Methods

fn prepare_render<'life0, 'life1, 'async_trait>(
&'life0 mut self,
state: &'life1 mut UIState<D>,
chunk: Rect
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Resizes the widget before the rendering step. May require changing the UIState - for example, if more books need to be fetched.

fn render_into_frame(&self, f: &mut Frame<'_, B>, state: &UIState<D>, chunk: Rect)

Renders the widget into the frame, using the provided space.

Arguments

f - A frame to render into.

chunk - A chunk to specify the size of the widget.

fn handle_input<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
event: Event,
state: &'life1 mut UIState<D>,
app: &'life2 mut AppChannel<D>
) -> Pin<Box<dyn Future<Output = Result<ApplicationTask, TuiError<D::Error>>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Processes event and may modify the UI state state, or the overall application state app, depending on specific event.

Example: If event triggers submission of command which adds books, then state may register new books, and app may be updated as new books are added.

Returns ApplicationTask::DoNothing in the event that the input is not captured.